Doxygen Changes v1
Doxygen comments YAY

+++ SpotBot/src/shares.h
@@ -1,9 +1,11 @@
 /** @file shares.h
  *  This file contains extern declarations of shares and queues which are used
- *  in more than one file of a demonstration project.
+ *  in more than one file of SpotBot project.
  * 
  *  @author JR Ridgely
+ *  @author Christian Clephan
  *  @date   2021-Oct-23 Original file
+ *  @date   2022-Nov-25 Edited for purposes of SpotBot project.
  *  @copyright (c) 2021 by JR Ridgely, released under the LGPL 3.0. 
  */
 
@@ -25,11 +27,13 @@ extern Share<bool> send_data;
 // A queue which triggers a task to print the count at certain times
 extern Queue<float> vel_queue;
 
-extern Queue<uint16_t> timer;
+// A queue which holds time values to be displayed by task_webserver
+extern Queue<float> timer;
 
-// A queue which holds velocity values over 1 rep
+// A queue which holds right IMU velocity values to be displayed by task_webserver
 extern Queue<float> vel_r;
 
+// A queue which holds left IMU velocity values to be displayed by task_webserver
 extern Queue<float> vel_l;
 
 #endif // _SHARES_H_
